+++ app/src/main/java/leonard/bakingapp/RecipeActivity.java
@@ -1,12 +1,10 @@
 package leonard.bakingapp;
 
 import android.content.ContentValues;
-import android.content.Context;
 import android.content.res.Configuration;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.AsyncTask;
-import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.GridLayoutManager;
@@ -15,11 +13,9 @@ import android.support.v7.widget.RecyclerView;
 import android.util.DisplayMetrics;
 import android.util.Log;
 
-import com.facebook.stetho.Stetho;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
-import org.json.JSONObject;
 //import org.json.JSONObject;
 //import org.json.simple.parser.JSONParser;
 //import org.json.simple.parser.ParseException;
@@ -36,8 +32,7 @@ import java.util.Collection;
 import java.util.List;
 
 import leonard.bakingapp.classes.Recipe;
-import leonard.bakingapp.data.RecipeDb;
-import leonard.bakingapp.data.RecipeTable;
+import leonard.bakingapp.database.RecipeTable;
 //import leonard.bakingapp.data.RecipeTable;
 
 public class RecipeActivity extends AppCompatActivity {
@@ -47,12 +42,7 @@ public class RecipeActivity extends AppCompatActivity {
     private RecyclerView.LayoutManager mLayoutManager;
     private ArrayList<Recipe> mRecipeArray;
 
-//    private static final String RECIPE_LIST_STATE = "recipeListStateKey";
     private static final String RECIPE_LIST = "recipeListKey";
-//    private String recipeJsonStr;
-
-    @Nullable
-//    private SimpleIdlingResource mSimpleIdlingResource;
 
 
     @Override
@@ -60,9 +50,6 @@ public class RecipeActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         mRecyclerView = findViewById(R.id.recipe_recycler_view);
-
-//        recipeJsonStr = "";
-//        mRecyclerView.setHasFixedSize(true);
 
         //change to grid in case of wide/tablet screen size
         int orientation = getResources().getConfiguration().orientation;
@@ -73,25 +60,10 @@ public class RecipeActivity extends AppCompatActivity {
         } else {
             mLayoutManager = new LinearLayoutManager(this);
         }
-        mRecyclerView.setLayoutManager(mLayoutManager);
 
+        mRecyclerView.setLayoutManager(mLayoutManager);
         mRecipeArray = new ArrayList<>();
 
-//        try {
-//            JSONObject recipeJson = new JSONObject(recipeJsonStr);
-//            JSONArray recipeJsonArray = recipeJson.getJSONArray("recipes");
-//            int jsonLength = recipeJsonArray.length();
-//            for(int i = 0; i < jsonLength; i++){
-//                mRecipeArray.add(recipeJsonArray.getJSONObject(i).getString("name"));
-//            }
-//        } catch (JSONException e) {
-//            e.printStackTrace();
-//        }
-
-//        mRecipeArray.add("Nutella Pie");
-//        mRecipeArray.add("Brownies");
-//        mRecipeArray.add("Yellow Cake");
-//        mRecipeArray.add("Cheesecake");
         mAdapter = new RecipeAdapter(mRecipeArray,this);
         mRecyclerView.setAdapter(mAdapter);
         mAdapter.notifyDataSetChanged();
@@ -103,7 +75,7 @@ public class RecipeActivity extends AppCompatActivity {
             updateRecipeCards(savedInstanceState.<Recipe>getParcelableArrayList(RECIPE_LIST));
         }
 
-//                 /* "ctrl+/" on this line to turn on Stetho debugger
+                 /* "ctrl+/" on this line to turn on Stetho debugger
         final Context context = this;
         Stetho.initialize(
                 Stetho.newInitializerBuilder(context)
@@ -113,8 +85,11 @@ public class RecipeActivity extends AppCompatActivity {
 
     }
 
+    /**
+     * Add recipes to database
+     * @param mRecipeArray The array that contains recipes to add to database
+     */
     private void addToRecipeDb(ArrayList<Recipe> mRecipeArray) {
-//        RecipeDb recipeDb = new RecipeDb();
 
         if(mRecipeArray!=null){
             for(int i = 0; i < mRecipeArray.size() ; i++)
@@ -132,11 +107,15 @@ public class RecipeActivity extends AppCompatActivity {
 
 
         }
-//        getContentResolver().insert(RecipeTable.CONTENT_URI,RecipeTable.getContentValues(recipeDb,false));
     }
 
+    /**
+     * Check to see if a recipe is already in the database
+     * @param recipe The recipe to be tested against the database
+     * @return True if recipe already exists in database
+     */
     private boolean checkIfExists(Recipe recipe) {
-//        TODO query to see if the recipe is already in the database
+        //query to see if the recipe is already in the database
         Boolean exists = false;
         Cursor cursor = getContentResolver().query(RecipeTable.CONTENT_URI,null,"col_recipe=\"" + recipe.name + "\"",null,null);
         if(cursor.getCount()>0){
@@ -146,33 +125,21 @@ public class RecipeActivity extends AppCompatActivity {
         return exists;
     }
 
-    private List<Recipe> extractRecipeNames(String recipeStr //JSONObject[] recipeJsonArray
-    ){
-//        List<String> recipeList = new ArrayList<>();
+    /**
+     * takes recipes from a Json and converts it to a list
+     * @param recipeJsonStr Json string that contains recipes that needs to be converted to List
+     * @return recipes in a list
+     */
+    private List<Recipe> extractRecipeNames(String recipeJsonStr){
         Gson gson = new Gson();
         Type collectionType = new TypeToken<Collection<Recipe>>(){}.getType();
-        List<Recipe> recipes = gson.fromJson(recipeStr,collectionType);
-        //            JSONArray recipeJson = new JSONArray(recipeStr);
-//            JSONArray recipeJsonArray = recipe;//jsonObject.getJSONArray("recipes");
-//        int jsonLength = recipes.length;//recipeJson.length();//recipeJsonArray.length;//.length();
-//        Log.d(TAG, "extractRecipeNames: jsonlength = " + jsonLength);
-//        for(int i = 0; i < jsonLength; i++){
-////                Log.d(TAG, "adding: " + recipeJsonArray[i].getString("name"));
-//            String str = recipes[i].name;
-////            recipeList.add(str);
-//        }
-//        Recipe recipe = recipes[0];
-//        Log.d(TAG, "extractRecipeNames: " + recipe.name);
+        List<Recipe> recipes = gson.fromJson(recipeJsonStr,collectionType);
         return recipes;
     }
 
-//    public android.support.test.espresso.IdlingResource getIdlingResource() {
-//        if (mSimpleIdlingResource == null){
-//            mSimpleIdlingResource = new SimpleIdlingResource();
-//        }
-//        return mSimpleIdlingResource;
-//    }
-
+    /**
+     * async task to retrieve the recipe Json from the URL
+     */
     public class FetchRecipesTask extends AsyncTask<String,Void,List<Recipe>>{
         @Override
         protected List<Recipe> doInBackground(String... strings) {
@@ -183,10 +150,8 @@ public class RecipeActivity extends AppCompatActivity {
 
             // Will contain the raw JSON response as a string.
             String recipeJsonStr;
-            JSONObject[] jsonArray = null;
 
             try {
-//                Log.d(TAG, "doInBackground: (1) trying");
                 Uri.Builder builder = new Uri.Builder();
                 builder.scheme("http")
                         .authority("d17h27t6h515a5.cloudfront.net")
@@ -197,7 +162,6 @@ public class RecipeActivity extends AppCompatActivity {
                         .appendEncodedPath("baking.json");
                 URL url = new URL(builder.build().toString());
 
-//                Log.d(TAG, "doInBackground: (2) connecting");
                 urlConnection = (HttpURLConnection) url.openConnection();
                 urlConnection.setRequestMethod("GET");
                 urlConnection.connect();
@@ -205,46 +169,23 @@ public class RecipeActivity extends AppCompatActivity {
                 // Read the input stream into a String
                 InputStream inputStream = urlConnection.getInputStream();
 
-//                JSONParser jsonParser = new JSONParser();
-//                jsonArray = (JSONArray) jsonParser.parse(
-//                        new InputStreamReader(inputStream, "UTF-8"));
-
 
                 StringBuffer stringBuffer = new StringBuffer();
                 if (inputStream == null) {
                     // Nothing to do.
                     return null;
                 }
-//
-//                Log.d(TAG, "doInBackground: (3) buffering");
+
                 reader = new BufferedReader(new InputStreamReader(inputStream));
-//                Gson gson = new Gson();
-////                Type collectionType = new TypeToken<Collection<JSONObject>>(){}.getType();
-////                jsonArray = gson.fromJson(reader, collectionType);
-//                jsonArray = gson.fromJson(reader,JSONObject[].class);
-//                Log.d(TAG, "doInBackground: " + jsonArray);
-////                Recipe recipe = gson.fromJson(jsonArray[0],Recipe.class);
-//                Log.d(TAG, "doInBackground: (4) lining");
 
                 String line;
                 while ((line = reader.readLine()) != null) {
-                    // Since it's JSON, adding a newline isn't necessary (it won't affect parsing)
-                    // But it does make debugging a *lot* easier if you print out the completed
-                    // buffer for debugging.
-//                    Log.d(TAG, "doInBackground: building");
                     stringBuffer.append(line + "\n");
                 }
-//                Log.d(TAG, "doInBackground: (5) lined");
 
-//                if (stringBuffer.length() == 0) {
-//                    // Stream was empty.  No point in parsing.
-//                    return null;
-//                }
-                recipeJsonStr = //gson.fromJson(reader,String.class);
-                stringBuffer.toString();
-//                Log.d(TAG, "doInBackground: stringed");
+                recipeJsonStr = stringBuffer.toString();
             } catch (IOException e) {
-                Log.d(TAG, "doInBackground: IOexception");
+                Log.e(TAG, "doInBackground: IOexception");
                 e.printStackTrace();
                 return null;
             } finally {
@@ -261,7 +202,6 @@ public class RecipeActivity extends AppCompatActivity {
             }
 
             return extractRecipeNames(recipeJsonStr);
-//            return null;
         }
 
         @Override
@@ -271,22 +211,22 @@ public class RecipeActivity extends AppCompatActivity {
         }
     }
 
+    /**
+     *  update the recipe arrays, database, and recyclerView
+     * @param recipes recipes to display and save
+     */
     private void updateRecipeCards(List<Recipe> recipes){
         if (recipes != null) {
             mRecipeArray.clear();
             mRecipeArray.addAll(recipes);
-//            mRecipeArray = recipes.clone();
         }
         addToRecipeDb(mRecipeArray);
-//        String name = mRecipeArray.get(0).name;
-//        mRecyclerView.setAdapter(mAdapter);
         mAdapter.notifyDataSetChanged();
     }
 
     @Override
     protected void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
-//        outState.putParcelable(RECIPE_LIST_STATE, mLayoutManager.onSaveInstanceState());
         outState.putParcelableArrayList(RECIPE_LIST, mRecipeArray);
     }
 }
