+++ app/src/main/java/leonard/bakingapp/StepsActivity.java
@@ -16,7 +16,7 @@ import java.util.List;
 import leonard.bakingapp.classes.Step;
 
 public class StepsActivity extends AppCompatActivity {
-    private static final String TAG = StepsActivity.class.getSimpleName();
+//    private static final String TAG = StepsActivity.class.getSimpleName();
     StepsPagerAdapter mStepsPagerAdapter;
     ViewPager mViewPager;
     private FloatingActionButton rightFAB, leftFAB;
@@ -74,18 +74,12 @@ public class StepsActivity extends AppCompatActivity {
 
             @Override
             public void onPageSelected(int position) {
-                //TODO initialize and set exoplayer to exoplayer view. Use only one instance of exoplayer to increase performance
-                Log.d(TAG, "onPageSelected: change to " + position);
+                //TODO initialize and set exoplayer to exoplayer view-- use only one instance of exoplayer to increase performance
 
 //                StepsFragment stepsFragment = (StepsFragment)mStepsPagerAdapter.getItem(position);
 //                stepsFragment.initializePlayer();
 //                stepsFragment.initializePlayer(mSimpleExoPlayer);
 
-//                String tag = "android:switcher:" + R.id.step_pager + ":" + position;
-//                Log.d(TAG, "tag: " + tag);
-//                List<Fragment> fragment = getSupportFragmentManager().getFragments();
-//                fragment.get(0);
-////                stepsFragment.initializePlayer();
 //                SimpleExoPlayerView exoPlayerView = stepsFragment.getView().findViewById(R.id.step_exoPlayerView);
 //                StepsFragment stepsFragment = (StepsFragment) mStepsPagerAdapter.getCurrentFragment();
 //                SimpleExoPlayerView exoPlayerView = stepsFragment.getView().findViewById(R.id.step_exoPlayerView);
@@ -104,10 +98,12 @@ public class StepsActivity extends AppCompatActivity {
         //begin with the selected step
         mViewPager.setCurrentItem(position);
 
-        //set
-
     }
 
+    /**
+     * high visibility of the navigation buttons according to position in the viewpager
+     * @param position  position in the viewpager
+     */
     private void updateNavButtons(int position){
         if(isLandscape){
             if(position==0){
@@ -123,13 +119,6 @@ public class StepsActivity extends AppCompatActivity {
         }
     }
 
-
-//    @Override
-//    protected void onDestroy() {
-//        super.onDestroy();
-//        Log.d(TAG, "onDestroy");
-//    }
-
     public void nextFabButton(View view) {
         mViewPager.setCurrentItem(mViewPager.getCurrentItem() + 1);
     }
@@ -144,12 +133,4 @@ public class StepsActivity extends AppCompatActivity {
 ////        stepsFragment.getPlayerView();
 //    }
 
-//    private List<Step> getStepArrayList(){
-//        return new ArrayList<Step>(){{
-//            add(new Step("Short Desc 01", "description", "https://d17h27t6h515a5.cloudfront.net/topher/2017/April/58ffd974_-intro-creampie/-intro-creampie.mp4", null));
-//            add(new Step("Short Desc 02", "description", "https://d17h27t6h515a5.cloudfront.net/topher/2017/April/58ffd9a6_2-mix-sugar-crackers-creampie/2-mix-sugar-crackers-creampie.mp4", null));
-//            add(new Step("Short Desc 03", "description", "https://d17h27t6h515a5.cloudfront.net/topher/2017/April/58ffd9a6_2-mix-sugar-crackers-creampie/2-mix-sugar-crackers-creampie.mp4", null));
-//            add(new Step("Short Desc 04", "description", "URL", null));
-//        }};
-//    }
 }
