+++ app/src/main/res/xml/recipe_widget_provider_info.xml
@@ -2,7 +2,7 @@
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
     android:initialKeyguardLayout="@layout/widget_page"
     android:initialLayout="@layout/widget_page"
-    android:minHeight="120dp"
+    android:minHeight="60dp"
     android:minWidth="180dp"
     android:previewImage="@drawable/example_appwidget_preview"
     android:resizeMode="horizontal|vertical"
